+++ madsim/src/sim/task.rs
@@ -274,6 +274,20 @@ impl TaskHandle {
         let mut nodes = self.nodes.lock();
         let node = nodes.get_mut(&id).expect("node not found");
         node.paused.clear();
+        node.info.killed.store(true, Ordering::Relaxed);
+        node.info.wakers.lock().drain(..).for_each(Waker::wake);
+
+        for sim in self.sims.lock().values() {
+            sim.reset_node(id);
+        }
+    }
+
+    /// Kill all tasks of the node and restart the initial task.
+    pub fn restart(&self, id: impl ToNodeId) {
+        debug!(node = %id, "restart");
+        let id = id.to_node_id(self);
+        let mut nodes = self.nodes.lock();
+        let node = nodes.get_mut(&id).expect("node not found");
         let new_info = Arc::new(NodeInfo {
             id,
             name: node.info.name.clone(),
@@ -286,21 +300,10 @@ impl TaskHandle {
             ctrl_c: Mutex::new(None),
         });
         let old_info = std::mem::replace(&mut node.info, new_info);
-        old_info.killed.store(true, Ordering::SeqCst);
+        node.paused.clear();
+        old_info.killed.store(true, Ordering::Relaxed);
         old_info.wakers.lock().drain(..).for_each(Waker::wake);
 
-        for sim in self.sims.lock().values() {
-            sim.reset_node(id);
-        }
-    }
-
-    /// Kill all tasks of the node and restart the initial task.
-    pub fn restart(&self, id: impl ToNodeId) {
-        debug!(node = %id, "restart");
-        let id = id.to_node_id(self);
-        self.kill_id(id);
-        let nodes = self.nodes.lock();
-        let node = nodes.get(&id).expect("node not found");
         if let Some(init) = &node.init {
             init(&Spawner {
                 sender: self.sender.clone(),
@@ -347,6 +350,14 @@ impl TaskHandle {
         // "ctrl-c" has never been called. kill node
         debug!(node = %id, "killed by ctrl-c");
         self.kill_id(id);
+    }
+
+    /// Returns whether the node is killed or exited.
+    pub fn is_exit(&self, id: impl ToNodeId) -> bool {
+        let id = id.to_node_id(self);
+        let nodes = self.nodes.lock();
+        let node = nodes.get(&id).expect("node not found");
+        node.info.killed.load(Ordering::Relaxed)
     }
 
     /// Create a new node.
@@ -512,6 +523,14 @@ impl Spawner {
             id,
             task: Mutex::new(Some(task.fallible())),
         }
+    }
+
+    /// Exit the current process (node).
+    pub(crate) fn exit(&self) {
+        debug!(node = %self.info.id, "exit");
+        // FIXME: clear paused tasks
+        self.info.killed.store(true, Ordering::Relaxed);
+        self.info.wakers.lock().drain(..).for_each(Waker::wake);
     }
 }
 
@@ -843,6 +862,7 @@ mod tests {
             assert_eq!(flag2.load(Ordering::Relaxed), 2);
             Handle::current().kill(node1.id());
             Handle::current().kill(node1.id());
+            assert!(Handle::current().is_exit(node1.id()));
 
             time::sleep_until(t0 + Duration::from_secs(5)).await;
             assert_eq!(flag1.load(Ordering::Relaxed), 2);
@@ -879,6 +899,7 @@ mod tests {
             assert_eq!(flag.load(Ordering::Relaxed), 2);
             Handle::current().kill(node.id());
             Handle::current().restart(node.id());
+            assert!(!Handle::current().is_exit(node.id()));
 
             time::sleep_until(t0 + Duration::from_secs(6)).await;
             assert_eq!(flag.load(Ordering::Relaxed), 2);
@@ -1033,6 +1054,37 @@ mod tests {
             handle.abort();
             let err = handle.await.unwrap_err();
             assert!(err.is_cancelled());
+        });
+    }
+
+    #[test]
+    fn exited() {
+        let runtime = Runtime::new();
+
+        let flag = Arc::new(AtomicUsize::new(0));
+        let flag_ = flag.clone();
+        let node = runtime
+            .create_node()
+            .init(move || {
+                let flag_ = flag_.clone();
+                async move {
+                    crate::task::spawn(async move {
+                        loop {
+                            time::sleep(Duration::from_secs(2)).await;
+                            flag_.fetch_add(2, Ordering::Relaxed);
+                        }
+                    });
+                    time::sleep(Duration::from_secs(5)).await;
+                    // exit here. the spawned task will be killed
+                }
+            })
+            .build();
+
+        runtime.block_on(async move {
+            assert!(!Handle::current().is_exit(node.id()));
+            time::sleep(Duration::from_secs(10)).await;
+            assert!(Handle::current().is_exit(node.id()));
+            assert_eq!(flag.load(Ordering::Relaxed), 4);
         });
     }
 }